+++ app/services/order/api/v1/communications/requests.go
@@ -1,4 +1,4 @@
-package orderapi
+package communications
 
 // OrderFilters struct lets user determine which orders they want to see.
 type OrderFilters struct {
